Replaced incremental recursive functions
Replaced recursive functions relying on an incrementing integer being passed with a map over the key shift integers in one case, and a fold in the other.

This removes the possibility and need to handle someone kicking off the recursion with a number they shouldn't (too high, too low), and makes it more stack safe.